Create compiler program which invokes Closure
The plan is to invoke Closure via the API, so we can keep a copy of the
process hot, and hopefully reduce the compilation time.

This adds an IntelliJ IDEA project and a Bazel build system to build the
compiler. You can invoke the compiler with the command:

    bazel run //java:compiler

This will run Closure, with the correct inputs and settings, and print
the minified output.