audio: Correctly set filter Q
The Q parameter on a biquad filter does not represent the Q factor, but
instead the peak filter gain (for lowpass and highpass filters). Added a
new integer scale encoding so dB values can be passed in as integers.

The lead sound is now softer, which I had been trying to do for a while.