Make game playable with gamepad
The player can move boxes with the gamepad. To start using the gamepad,
press the action button on the gamepad. This will switch control from
keyboard control to gamepad control. Control can return to the keyboard
or change to a different gamepad by pressing a button.

This is a major input system refactor. The keyboard and each gamepad
object is assigned to a controller, rather than using a global table for
all input devices.