Relax restrictions on joystick input for pushing
Previously, the joystick had to point within a certain margin of a
cardinal direction (within a 1:2 slope). This was intended to make
accidental joystick movements not register, but instead, it just made it
harder to purposefully push with the joystick. For accidental pushes, it
is enough that the player can abort the push by letting go before the
box moves past a threshold. The fact that the joystick is pushed past a
certain distance makes it likely that the push was intentional to begin
wiht.